refactor(auto_source): decompose JsonState scraper into modular peer subcomponents
Decompose monolithic JsonState scraper into focused peer subcomponents:
- DocumentScanner: script tag discovery and globals extraction
- ValueFinder: recursive key matching and nested structure navigation
- CandidateDetector: article array detection heuristics
- ArticleNormalizer: article hash transformation and URL normalization

Also adds unit tests for ValueFinder.
